requirements.txt
pyzk==0.9.0
PyYAML==6.0.1
schedule==1.2.0
pywin32==306
psutil==5.9.5
colorama==0.4.6
pytest==7.4.2 